Julia Reference

Sources

DatasetManager.AbstractSource — Type

AbstractSource

The abstract supertype for custom data sources. Implement a subtype of AbstractSource to enable loading your data while benefitting from multiple-dispatch.

Extended help

All subtypes of AbstractSource must:

have a path field or extend the sourcepath function.

accept a single string, the absolute path of the source file, in the constructor.

Example

struct MySource <: AbstractSource
    path::String
end

DatasetManager.readsource(src::MySource) = customreadfunc(sourcepath(src))
DatasetManager.readsegment(seg::Segment{MySource}) = readsource(seg.source)[start:finish]

DatasetManager.sourcepath — Function

sourcepath(src) -> String

Return the absolute path to the src.

DatasetManager.readsource — Function

readsource(src::S; kwargs...) where {S <: AbstractSource}

Read the source data from storage.

Trials

DatasetManager.DataSubset — Type

DataSubset(name, source::Type{<:AbstractSource}, dir, pattern)

Describes a subset of data, where files found within dir, with (absolute) paths which match pattern (using glob syntax), are all of the same AbstractSource subtype.

Examples

julia> DataSubset("events", EventsSource, "path/to/subset", "Subject [0-9]*/events/*.tsv")
DataSubset("events", EventsSource, "path/to/subset", "Subject [0-9]*/events/*.tsv")

DatasetManager.TrialConditions — Type

TrialConditions(conditions, labels; <keyword arguments>)

Describes the experimental conditions and the labels for levels within each condition.

Arguments

conditions is a collection of condition names (eg (:medication, :strength))

labels is a Dict with keys for each condition name (eg haskey(labels, :medication)).

Each key gets a collection of the labels for all levels and any transformation desired for that condition.

Keyword arguments

required=conditions: The conditions which every trial must have (in the case of some

trials having optional/additional conditions).

types=fill(String, length(conditions): The (Julia) types for each condition (eg

[String, Int])

sep="[_-]": The character separating condition labels

DatasetManager.Trial — Type

Trial(subject, name, [conditions[, sources]])

Describes a single trial, including a reference to the subject, trial name, trial conditions, and relevant sources of data.

DatasetManager.findtrials — Function

findtrials(subsets::AbstractVector{DataSubset}, conditions::TrialConditions;
    <keyword arguments>) -> Vector{Trial}

Find all the trials matching conditions which can be found in subsets.

Keyword arguments:

subject_fmt=r"(?<=Subject )(?<subject>\d+)": The format that the subject identifier will appear in file paths.

ignorefiles::Union{Nothing, Vector{String}}=nothing: A list of files, given in the form of an absolute path, that are in any of the subsets folders which are to be ignored.

defaultconds::Union{Nothing, Dict{Symbol}}=nothing: Any conditions which have a default level if the condition is not found in the file path.

Segments

DatasetManager.Segment — Type

Segment(trial, source::Union{AbstractSource,String}; [start, finish, conditions])

Describes a portion of a source in trial from time start to finish with segment specific conditions, if applicable.

If source isa String, it must refer to a source that exists in trial. If the start is omitted, the segment will start at the beginning of the source/trial. If the finish is omitted, the segment will be from time start to the end of the source/trial.

Any conditions present in trial will be merged into the conditions for the segment.

Example:

t = Trial(1, "intervention")

struct MySource <: AbstractSource
    path::String
end

seg = Segment(t, MySource; start=0.0, finish=10.0, conditions=Dict(:group => "control"))

# Use a source that already exists in the trial
seg5 = Segment(t, "main")

DatasetManager.SegmentResult — Type

SegmentResult(segment::Segment, results::Dict{Symbol)

Contains the results of any analysis/analyses performed on the trial segment in a Dict.

Example:

segresult = SegmentResult(seg, Dict(:avg => 3.5))

DatasetManager.readsegment — Function

readsegment(seg::Segment; kwargs...)

Return the portion of seg.source from seg.start to seg.finish.

Extended help

Subtypes of AbstractSource must implement this function to enable reading Segment{MySource} with this function.

DatasetManager.trial — Function

trial(seg::Union{Segment,SegmentResults}) -> Trial

Return the parent Trial for the given Segment or SegmentResult

DatasetManager.subject — Function

subject(seg::Union{Segment,SegmentResults})

Return the subject reference for the parent Trial of the given Segment or SegmentResult

DatasetManager.conditions — Function

conditions(seg::Union{Segment,SegmentResults}) -> Dict{Symbol}

Return the conditions for the given Segment or SegmentResult