Julia Reference

Sources

DatasetManager.AbstractSource — Type

AbstractSource

The abstract supertype for custom data sources. Implement a subtype of AbstractSource if Source{S} is not sufficient for your source requirements (e.g. your source has additional information besides the path, such as encoding parameters, compression indicator, etc, that needs to be associated with each instance).

Extended help

All subtypes of AbstractSource must:

have a path field or extend the sourcepath function.

accept a single string, the absolute path of the source file, in the constructor.

Minimum method definitions necessary for complete functionality of a subtype of AbstractSource (i.e. Source{S} or otherwise):

readsource

sourcepath (if defining a new subtype of AbstractSource)

Additional methods to improve user experience and/or additional functionality

readsegment

generatesource (if enabling requiresource! generation)

dependencies (if defining a generatesource method)

srcext

srcname_default

Example

struct MySource <: AbstractSource
    path::String
end

DatasetManager.readsource(src::MySource) = customreadfunc(sourcepath(src))
DatasetManager.readsegment(seg::Segment{MySource}) = readsource(seg.source)[start:finish]

DatasetManager.Source — Type

Source{S}(path)

A basic source, where S can be any singleton or existing type

DatasetManager.readsource — Function

readsource(src::S; kwargs...) where {S <: AbstractSource}

Read the source data from file.

DatasetManager.sourcepath — Function

sourcepath(src) -> String

Return the absolute path to the src.

DatasetManager.requiresource! — Function

requiresource!(trial, src::Union{S,Type{S}}; force=false, deps, kwargs...) -> nothing
requiresource!(trial, name::String => src::Union{S,Type{S}}; force=false, deps, kwargs...)

Require a source src in trial, generate src if not present, or throw an error if src is not present and cannot be generated. Returns nothing. Keyword argument deps defaults to dependencies(src), but can be manually specified as a name => src::AbstractSource pair per dependency if there exist multiple sources that would match a dependent source type. kwargs... are passed on to the generatesource function for src.

Keyword arguments

force=false: Force generating src, even if it already exists

deps=dependencies(src): Manually specify particular dependencies

Examples

julia> dependencies(Source{Events})
(Source{C3DFile},)

julia> requiresource!(trial, Source{Events})

julia> requiresource!(trial, Source{Events}; force=true, deps=("mainc3d" => Source{C3DFile}))

DatasetManager.generatesource — Function

generatesource(trial, src::Union{S,Type{S}}, deps; kwargs...) where S <: AbstractSource -> newsrc::typeof(src)

Generate source src using dependent sources deps from trial. Returns a source of the same type as src, but is not required to be exactly equal to src (i.e. a different sourcepath(newsrc) is acceptable).

DatasetManager.dependencies — Function

dependencies(src::Union{S,Type{S}})

Get the sources that src depends on to be generated by generatesource.

DatasetManager.srcext — Function

srcext(src::Union{S,Type{S}})

Actual file extension or default file extension for a `src` or `src` type. Period (.) should be the first letter.

DatasetManager.srcname_default — Function

Default name for a source in Trial.sources

Trials

DatasetManager.DataSubset — Type

DataSubset(name, source::Type{<:AbstractSource}, dir, pattern)

Describes a subset of data, where files found within dir, with (absolute) paths which match pattern (using glob syntax), are all of the same AbstractSource subtype.

Examples

julia> DataSubset("events", Source{Events}, "path/to/subset", "Subject [0-9]*/events/*.tsv")
DataSubset("events", Source{Events}, "path/to/subset", "Subject [0-9]*/events/*.tsv")

DatasetManager.TrialConditions — Type

TrialConditions(conditions, labels; <keyword arguments>)

Describes the experimental conditions and the labels for levels within each condition.

Arguments

conditions is a collection of condition names (eg (:medication, :strength))

labels is a Dict with keys for each condition name (eg haskey(labels, :medication)).

Each key gets a collection of the labels for all levels and any transformation desired for that condition.

Keyword arguments

required=conditions: The conditions which every trial must have (in the case of some

trials having optional/additional conditions).

types=fill(String, length(conditions): The (Julia) types for each condition (eg

[String, Int])

sep="[_-]": The character separating condition labels

DatasetManager.Trial — Type

Trial(subject, name, [conditions, sources])

Describes a single trial, including a reference to the subject, trial name, trial conditions, and relevant sources of data.

DatasetManager.hassource — Function

hassource(trial, src::Union{String,S<:AbstractSource}) -> Bool

Check if trial has a source with key or type src.

Examples

julia> hassource(trial, "model")
false

julia> hassource(trial, Source{Events})
true

DatasetManager.getsource — Function

getsource(trial, src::Union{String,Type{<:AbstractSource}}) -> <:AbstractSource
getsource(trial, name::String => src::Type{<:AbstractSource}) -> <:AbstractSource

Return a source from trial with key or type src. When the second argument is a pair, a source with key name will returned, or of type src if no source name is present.

If multiple sources of type src are present, the desired source must be accessed by name/key only or an error will be thrown.

DatasetManager.sources — Function

sources(trial::Trial{ID}) -> Dict{String,AbstractSource}

Get the sources for trial

DatasetManager.findtrials — Function

findtrials(subsets::AbstractVector{DataSubset}, conditions::TrialConditions;
    <keyword arguments>) -> Vector{Trial}

Find all the trials matching conditions which can be found in subsets.

Keyword arguments:

subject_fmt=r"(?<=Subject )(?<subject>\d+)": The format that the subject identifier will appear in file paths.

ignorefiles::Union{Nothing, Vector{String}}=nothing: A list of files, given in the form of an absolute path, that are in any of the subsets folders which are to be ignored.

defaultconds::Union{Nothing, Dict{Symbol}}=nothing: Any conditions which have a default level if the condition is not found in the file path.

debug=false: Show Regex and files that did not match for each subset. Use to debug TrialConditions definitions or issues with subject_fmt failing to match subject ID's.

DatasetManager.analyzedataset — Function

analyzedataset(f, trials, Type{<:AbstractSource}; kwargs...) -> Vector{SegmentResult}

Call function f on every trial in trials in parallel (multi-threaded). If f errors for a given trial, the SegmentResult for that trial will be empty (no results), and the trial and error will be shown after the analysis has finished.

Keyword arguments

threaded=true: Analyze trials using multiple threads

enable_progress=true: Enable the progress meter

DatasetManager.export_trials — Function

export_trials([f,] trials, dir[, sources])

Export (copy) sources in trials to outdir. When left unspecified, sources is set to all unique sources found in trials. Optionally can be given a function f, which must accept 2 arguments (a trial and a src which is of eltype(sources)), to control the names of the exported data. The default behavior exports all sources to dir with no subdirectories, using the naming schema "$trial.subject_$srcname_basename(sourcepath)" (pseudo-code).

Examples

julia> export_trials(trials, pwd()) do trial, source
    "$(subject(trial))_$(conditions(trial)[:group]).$(srcext(source))"
end

Segments

DatasetManager.Segment — Type

Segment(trial, src::Union{S,Type{S},String}; [start, finish, conditions])

Describes a portion of a source in trial from time start to finish with segment specific conditions, if applicable.

If src is a String or AbstractSource, it must refer to a source that exists in trial. If the start is omitted, the segment will start at the beginning of the source/trial. If the finish is omitted, the segment will be from time start to the end of the source/trial.

Any conditions present in trial will be merged into the conditions for the segment.

Example:

t = Trial(1, "intervention")

struct MySource <: AbstractSource
    path::String
end

seg = Segment(t, MySource; start=0.0, finish=10.0, conditions=Dict(:group => "control"))

# Use a source that already exists in the trial
seg5 = Segment(t, "main")

DatasetManager.SegmentResult — Type

SegmentResult(segment::Segment, results::Dict{Symbol)

Contains the results of any analysis/analyses performed on the trial segment in a Dict.

Example:

segresult = SegmentResult(seg, Dict(:avg => 3.5))

DatasetManager.readsegment — Function

readsegment(seg::Segment{S}; warn=true, kwargs...) where S <: AbstractSource

If defined for S, returns the portion of seg.source from seg.start to seg.finish. Otherwise, equivalent to readsource (i.e. no trimming of time-series occurs). Warns by default if the main method is called.

DatasetManager.trial — Function

trial(seg::Union{Segment,SegmentResult}) -> Trial

Return the parent Trial for the given Segment or SegmentResult

DatasetManager.subject — Function

subject(trial::Trial{ID}) -> subject::ID

Get the subject identifier for trial

subject(seg::Union{Segment,SegmentResult})

Return the subject reference for the parent Trial of the given Segment or SegmentResult

DatasetManager.segment — Function

Get the segment of a SegmentResult

DatasetManager.source — Function

source(seg::Union{Segment,SegmentResult}) -> AbstractSource

Return the source for the parent Trial of the given Segment or SegmentResult

DatasetManager.conditions — Function

conditions(trial::Trial{ID}) -> Dict{Symbol,Any}

Get the conditions for trial

conditions(seg::Union{Segment,SegmentResult}) -> Dict{Symbol}

Return the conditions for the given Segment or SegmentResult

DatasetManager.results — Function

Get the results of a SegmentResult

DatasetManager.resultsvariables — Function

resultsvariables(sr::Union{SegmentResult,Vector{SegmentResult}})

Get the unique variables for SegmentResults.

Utilities

DatasetManager.summarize — Function

summarize([io,] trials; verbosity=5)

Summarize a vector of Trials.

Summarizes:

Number of unique subjects and lists IDs

Total number of trials and trials per subject

Unique conditions and observed levels, unique combinations of conditions

Unique sources and source types

Examples

julia> summarize(trials)
Subjects:
 └ 15: 1  2  3  4  5  6  7  8  9  10  11  12  13  14  15
Trials:
 ├ Number of trials: 90
 └ Number of trials per subject:
   └ 6: 15/15 (100%)
Conditions:
 ├ Observed levels:
 │ ├ sym => ["asym", "sym"]
 │ └ arms => ["active", "held", "norm"]
 └ Unique level combinations observed: 6 (full factorial)
      sym │   arms │ # trials
    ──────┼────────┼──────────
     asym │ active │ 15
     asym │   held │ 15
     asym │   norm │ 15
      sym │ active │ 15
      ⋮   │   ⋮    │ ⋮
Sources:
 └ "events" => Source{Events}, 90 trials (100%)

DatasetManager.stack — Function

stack(rs::Vector{SegmentResult}, conds; [variables])

Compile the results into a stacked, long form DataFrame